+++ server-src/routes-enabled/50-accept.js
@@ -12,16 +12,15 @@ const acceptRoute = Express.Router();
 const dbi = new DBInterface().open();
 const emi = new EMInterface();
 
-acceptRoute.use(`/accept`, async (req, res, next) => {    
+acceptRoute.use(`/accept`, async (req, res, next) => {
     try {
         acceptRequest(req.query.hash, req.query.email, dbi, emi);
         res.redirect(`/status?hash=${req.query.hash}`);
-        logger.log(JSON.stringify({ ...req.query, action: "accept" }));
     } catch (error) {
         logger.error(error.toString());
         reject500(req, res);
     }
 });
 
-export default acceptRoute;
+export {acceptRoute as default, emi, dbi}