+++ test/integration/accept_request.js
@@ -4,7 +4,6 @@ import Server from "../../server-src/Server.js";
 import assert from "assert";
 import { options } from "../../server-src/parseArgs.js";
 import logger from "../../server-src/setupLogger.js";
-import { emi } from "../../server-src/routes-enabled/50-accept.js";
 
 options.flags.push({
     "long": "out",
@@ -40,7 +39,6 @@ try {
 if (FS.existsSync(args.in)) {
     const file = FS.readFileSync(args.in);
     await fetchSubmit(JSON.parse(file));
-    await emi.wait();
     server.stop();
 } else {
     console.log(`Input file not found: ${args.in}`);
@@ -63,8 +61,4 @@ async function fetchSubmit(inputFileData) {
     logger.console(`response written to '${args.out}'`);
     FS.writeFileSync(args.out, response);
     logger.verbose(response);
-
-    if (args.browse) {
-        exec(`google-chrome ${args.out}`);
-    }
 }